requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
python-dateutil==2.9.0
pytz==2024.1
filelock==3.14.0
